Move the shim-coverage check into the sign-exclusions test that owns it
windowsSignExclusions.test.ts already models app-builder-lib's
shouldSignFile faithfully; its only weakness was a hand-written shim list,
which is why it passed both with and without the fix for run 33391945572.
Put the lockfile-derived enumeration there rather than in a second file
asserting the same property with a weaker matcher, and drop the file added
in the previous commit.

Mutation-proven in its new home: removing the nested @puppeteer/browsers
entry from electron-builder.yml fails the coverage assertion.
